[BE] refactor: 나머지 리팩토링
